Append partition to broker topic stats
This change adds the partition id information to the broker topic stats
so the information is more actionable. Without the partition id, the
metric shows the combined value across all partitions of the same topic.

In a cluster environment, one broker could be hosting 10 partitions of
some topic while another broker could be hosting the other 20 partitions
of the same topic. The latter broker does not necessarily reflect an
unbalanced load compared to the former broker under such circumstance,
especially if the former is hosting another 10 unrelated but similar
load topic-partitions.