Introduce controlled.shutdown.max.under.replication
Summary:
Enabling controlled.shutdown.enable allows all follower brokers to be
shut down, as long as the leader broker is still alive.

This change introduces an additional config to limit the parallelism
of broker shutdowns within the same cluster. The config specifies the
number of replicas that can be safely shutdown at any moment. This flag
requires controlled.shutdown.enable to also be set.

This behavior allows sending a cluster wide shutdown to all brokers and
the controller will essentially perform a controlled rolling shutdown.

Reviewers: grk, vinoth, nharkins, mlai

Differential Revision: https://code.uberinternal.com/D86138