Introduce kafka.log.Log.Size.<topic>.<partition>.util
Summary:
There is no easy way to monitor disk utilization per topic-partition
across the entire cluster in a generalized way. This change introduces
a new metric which divides the current size of each topic-partition with
its corresponding log.retention.bytes. The value is a fraction <=1.0, as
oppose to an absolute byte size which may be different across topics.

Reviewers: grk

Reviewed By: grk

Differential Revision: https://code.uberinternal.com/D88305